+++ client/src/components/Events/banner1.js
@@ -1,0 +1,28 @@
+import React from 'react'
+
+function banner1() {
+    return (
+        <>
+            <div className="relative flex flex-col items-start mb-6 ml-20 pt-16 pb-[177px]">
+                <div
+                    className="w-3/5 h-[609px] bg-[url(https://file.rendit.io/n/ih5eyGEzbAMlyH5z0rji.png)] bg-cover bg-50%_50% bg-blend-normal bg-no-repeat absolute top-0 left-[605px] flex flex-col" id="IMG" >
+                    <div className="bg-[rgba(13,_4,_4,_0.54)] mr-[-132px] h-[609px] shrink-0" />
+                </div>
+                <div className="text-[128px] font-['Inter'] relative w-1/2 h-full" id="LETSPARTYFORJESUS3" >
+                    LET’S{" "}
+                    <div className="font-semibold contents" id="LETSPARTYFORJESUS">
+                        PARTY FO
+                    </div>
+                    <div className="font-semibold text-[#fffefe] contents" id="LETSPARTYFORJESUS1">
+                        R<br />
+                    </div>
+                    <div className="font-bold contents" id="LETSPARTYFORJESUS2">
+                        JESUS
+                    </div>
+                </div>
+            </div>
+        </>
+    )
+}
+
+export default banner1
