+++ client/src/components/Events/banner2.js
@@ -1,0 +1,27 @@
+import React from 'react';
+
+function Events() {
+  return (
+    <>
+      <div className="text-5xl font-['Inter'] font-bold self-end w-56 mb-10 mr-[599px]">
+        EVENTS
+      </div>
+      <div className="bg-[url(https://file.rendit.io/n/QzteZ71X4L0K9yHOzuEL.png)] bg-cover bg-50%_50% bg-blend-normal bg-no-repeat flex flex-row gap-1 h-[457px] shrink-0 items-center mb-24 ml-20 mr-16 pl-16 pr-12" id="OCWJW">
+        <div className="relative flex flex-col pt-px w-1/2">
+          <img src="https://file.rendit.io/n/0w3Yhx9ACanaF5vGgCiy.svg" alt='a' className="w-[548px] h-[307px] absolute top-0 left-px" />
+          <div className="bg-black/56.00000000000001 relative mr-2 h-[305px] shrink-0" />
+        </div>
+        <div className="flex flex-col gap-1 w-1/2">
+          <div className="whitespace-nowrap text-5xl font-['Inter'] font-black text-white ml-8 mr-10">
+            UPCOMING EVENTS
+          </div>
+          <div className="text-center text-4xl font-['Istok_Web'] tracking-[-2] text-white h-[59.22%]">
+            "Enriching faith through inspiring events
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}
+
+export default Events;
