--- client/src/components/Events/index.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Events() {
-  return (
-    <>
-      Events
-    </>
-  );
-}
-
-export default Events;
