+++ client/src/screens/SafinaPages/events.js
@@ -1,6 +1,9 @@
 import React from 'react'
 import Navbar from '../../components/Safina/navbar'
 import Footer from '../../components/Safina/footer'
+import Banner1 from '../../components/Events/banner1'
+import Banner2 from '../../components/Events/banner2'
+import Banner3 from '../../components/Events/banner3'
 
 
 function Events() {
@@ -10,121 +13,32 @@ function Events() {
             <div className="flex flex-col w-full" id="NewRootRoot">
                 <div className="overflow-hidden bg-white flex flex-col" id="Desktop">
                     <div className="bg-[url(https://file.rendit.io/n/q1wX6sn8kCWVylrrcci1.png)] bg-cover bg-50%_50% bg-blend-normal bg-no-repeat flex flex-col pt-6 pb-24" id="Image1" >
-                        <div className="bg-[#d9d9d9] h-12 shrink-0 mb-6 ml-10 mr-8 rounded-lg" />
-                        {/* The First Component */}
-                        <div className="relative flex flex-col items-start mb-6 ml-20 pt-16 pb-[177px]">
-                            <div
-                                className="w-3/5 h-[609px] bg-[url(https://file.rendit.io/n/ih5eyGEzbAMlyH5z0rji.png)] bg-cover bg-50%_50% bg-blend-normal bg-no-repeat absolute top-0 left-[605px] flex flex-col" id="IMG" >
-                                <div className="bg-[rgba(13,_4,_4,_0.54)] mr-[-132px] h-[609px] shrink-0" />
-                            </div>
-                            <div className="text-[128px] font-['Inter'] relative w-1/2 h-full" id="LETSPARTYFORJESUS3" >
-                                LET’S{" "}
-                                <div className="font-semibold contents" id="LETSPARTYFORJESUS">
-                                    PARTY FO
-                                </div>
-                                <div
-                                    className="font-semibold text-[#fffefe] contents"
-                                    id="LETSPARTYFORJESUS1">
-                                    R<br />
-                                </div>
-                                <div className="font-bold contents" id="LETSPARTYFORJESUS2">
-                                    JESUS
-                                </div>
-                            </div>
+                        <div className="bg-[#d9d9d9] h-12 shrink-0 mb-6 ml-10 mr-8 rounded-lg" >
+                            {/* Fix Text */}
                         </div>
-                        <div className="text-5xl font-['Inter'] font-bold self-end w-56 mb-10 mr-[599px]">
-                            EVENTS
-                        </div>
-                        <div
-                            className="bg-[url(https://file.rendit.io/n/QzteZ71X4L0K9yHOzuEL.png)] bg-cover bg-50%_50% bg-blend-normal bg-no-repeat flex flex-row gap-1 h-[457px] shrink-0 items-center mb-24 ml-20 mr-16 pl-16 pr-12" id="OCWJW">
-                            <div className="relative flex flex-col pt-px w-1/2">
-                                <img
-                                    src="https://file.rendit.io/n/0w3Yhx9ACanaF5vGgCiy.svg" alt='a' className="w-[548px] h-[307px] absolute top-0 left-px" />
-                                <div className="bg-black/56.00000000000001 relative mr-2 h-[305px] shrink-0" />
-                            </div>
-                            <div className="flex flex-col gap-1 w-1/2">
-                                <div className="whitespace-nowrap text-5xl font-['Inter'] font-black text-white ml-8 mr-10">
-                                    UPCOMING EVENTS
-                                </div>
-                                <div className="text-center text-4xl font-['Istok_Web'] tracking-[-2] text-white h-[59.22%]">
-                                    "Enriching faith through inspiring events
-                                </div>
-                            </div>
-                        </div>
-                        <div className="flex flex-row gap-1 items-end ml-20 mr-2">
-                            <div className="bg-[#d9d9d9] relative flex flex-col w-[364px] shrink-0 items-center mt-3 mr-24 pb-2 px-1">
-                                <div
-                                    className="whitespace-nowrap text-5xl font-['Inter'] font-semibold absolute top-16 left-20 h-32 w-1/2"
-                                    id="RISINGSTARS"
-                                >
-                                    RISING <br />
-                                    STARS
-                                </div>
-                                <img
-                                    src="https://file.rendit.io/n/ppNpcb4LnSG3mgB6av6X.png"
-                                    className="relative"
-                                    id="IMG1"
-                                    alt='Nice'
-                                />
-                            </div>
-                            {/* The slide */}
-                            <div className="self-start relative flex flex-col mr-24 w-[364px] shrink-0 items-center">
-                                <div className="whitespace-nowrap text-5xl font-['Inter'] font-semibold absolute top-[109px] left-12 h-16 w-[267px]">
-                                    MAIN EXP
-                                </div>
-                                <img
-                                    src="https://file.rendit.io/n/ub43FsWbzxdlU2CDjBTt.svg"
-                                    className="relative"
-                                    alt='Nice'
-                                />
-                            </div>
-                            <img
-                                src="https://file.rendit.io/n/UA9EZq4Qme2pYitmGtAH.svg"
-                                className=""
-                                alt='Nice'
-                            />
-                            <img
-                                src="https://file.rendit.io/n/MZzU7q5pMQbfZ2ggpTAB.svg"
-                                className="mb-20 w-12 shrink-0"
-                                id="Arrowrightsolid"
-                                alt='Nice'
-                            />
-                        </div>
-                        <div className="flex flex-row gap-48 items-center mb-12 ml-20 mr-32">
-                            <div className="text-center text-3xl font-['Istok_Web'] font-bold tracking-[-1.6] self-start mb-3 w-[346px] shrink-0 h-[81.54%]">
-                                join us in prayers in rising stars
-                            </div>
-                            <div className="text-center whitespace-nowrap text-3xl font-['Istok_Web'] font-bold tracking-[-1.6] mr-8 w-56 shrink-0">
-                                Learn the word,
-                            </div>
-                            <div className="text-center whitespace-nowrap text-3xl font-['Istok_Web'] font-bold tracking-[-1.6] self-end w-64 shrink-0">
-                                mid-week services
-                            </div>
-                        </div>
-                        <div className="self-center relative flex flex-col justify-center mb-12 w-[366px] pt-5 pb-3">
-                            <div className="w-[270px] h-20 bg-[#06123d] absolute top-0 left-12 rounded-[20px]" />
-                            <div className="text-center text-4xl font-['Inter'] font-semibold text-white relative h-full">
-                                VIEW MORE
-                            </div>
-                        </div>
-                        <div className="relative flex flex-col pb-1 mb-24 mx-[-16px]">
-                            <div className="w-[1472px] h-[400px] bg-[#d9d9d9] absolute top-2 left-0" />
-                            <img
-                                src="https://file.rendit.io/n/H6h4JWdrJZ9ti8YWu2j8.png"
-                                alt='ark church'
-                                className="w-[1440px] h-[412px] absolute top-1 left-4"
-                                id="OCWJW1" />
-                            <div className="bg-[rgba(24,_22,_64,_0.83)] relative flex flex-col h-[411px] shrink-0 items-end pt-40 pb-32 pr-[351px]">
-                                <div className="text-center whitespace-nowrap text-5xl font-['Inter'] font-semibold text-white w-3/5">
-                                    BIBLE VERSE
-                                    <br />
-                                    WITH BACKGROUND GIF
-                                </div>
-                            </div>
-                        </div>
+                        {/* The First Component Banner*/}
+
+                        <Banner1 />
+
+                        {/* End Component Banner */}
+                        {/* ------------------------------------------------------------------------------------------------ */}
+                        {/* The second Component Banner*/}
+
+                        <Banner2 />
+
+                        {/* The End second Component Banner*/}
+                        {/* ------------------------------------------------------------------------------------------------ */}
+                        {/* The  Third Component Banner*/}
+
+                        <Banner3 />
+
+                        {/* The End Third Component Banner*/}
+                        {/* ------------------------------------------------------------------------------------------------ */}
                     </div>
                 </div>
+                {/* The End Desktop */}
                 <Footer />
+                {/* The End Footer */}
             </div>
         </>
     )
